requirements.txt
disnake
python-dotenv
matplotlib
pillow